to-ts, fix some problems.
