+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "markdown-it-pangu",
   "version": "1.0.2",
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "description": "A markdown-it plugin which inserts whitespace into your markdown files.",
   "author": "Shigma <1700011071@pku.edu.cn>",
   "license": "MIT",
@@ -14,6 +14,7 @@
     "url": "https://github.com/Shigma/markdown-it-pangu.git"
   },
   "devDependencies": {
+    "@types/pangu": "^4.0.0",
     "@vuepress/plugin-back-to-top": "^1.0.0-alpha.44",
     "@vuepress/plugin-medium-zoom": "^1.0.0-alpha.44",
     "@vuepress/shared-utils": "^1.0.0-alpha.44",
