+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "outDir": "./build",
+        "allowJs": true,
+        "target": "es5",
+        "esModuleInterop": true,
+    },
+    "include": [
+        "./src/**/*"
+    ]
+}
